# services/auth/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API - Servicio de Autenticación (Gestor de Stock Juguetería)",
    "description": "Login, JWT de acceso/refresh, roles, y gestión de usuarios.",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "description": "Autentica con email/password y devuelve un par de tokens (access +\nrefresh). Bloquea temporalmente tras demasiados intentos fallidos\nseguidos (ver app/core/redis_client.py).",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Refresh",
        "description": "Intercambia un refresh token válido (y no usado antes) por un access\ntoken nuevo. El refresh token usado queda invalidado inmediatamente\n(rotación de un solo uso, issue #7): la respuesta no reutiliza el mismo\nrefresh token, siempre emite uno nuevo también.\n\nNota de diseño: se responde solo con el nuevo access_token en el body\npara mantener el schema simple; el nuevo refresh_token se agrega en un\nheader propio (X-New-Refresh-Token) para no romper el contrato de\nAccessTokenResponse. El frontend debe leerlo de ahí y reemplazar el\nrefresh token guardado.",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessTokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout",
        "description": "Revoca el access token actual (queda en blacklist) y, si se envía en\nel body, también el refresh token asociado, para cerrar la sesión por\ncompleto en ambos frentes.",
        "operationId": "logout_api_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RefreshRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Datos"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Quien Soy",
        "description": "Devuelve los datos del usuario autenticado actual.",
        "operationId": "quien_soy_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/usuarios": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Listar Usuarios",
        "description": "Lista todos los usuarios. Solo accesible por un admin.",
        "operationId": "listar_usuarios_api_v1_auth_usuarios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UsuarioOut"
                  },
                  "type": "array",
                  "title": "Response Listar Usuarios Api V1 Auth Usuarios Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Crear Usuario",
        "description": "Crea un usuario nuevo (admin o empleado). Solo accesible por un admin.",
        "operationId": "crear_usuario_api_v1_auth_usuarios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/": {
      "get": {
        "tags": [
          "Root"
        ],
        "summary": "Root",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "Root"
        ],
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AccessTokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "AccessTokenResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest"
      },
      "RolUsuario": {
        "type": "string",
        "enum": [
          "admin",
          "empleado"
        ],
        "title": "RolUsuario",
        "description": "Roles soportados en este proyecto (issue #7): dos roles, admin y\nempleado. admin tiene acceso a todo, incluida la consulta de auditoría;\nempleado puede operar productos/movimientos pero no ver la auditoría ni\ngestionar usuarios."
      },
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token"
        ],
        "title": "TokenResponse"
      },
      "UsuarioCreate": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 72,
            "minLength": 8,
            "title": "Password",
            "description": "Mínimo 8 caracteres."
          },
          "nombre": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Nombre"
          },
          "rol": {
            "$ref": "#/components/schemas/RolUsuario",
            "default": "empleado"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password",
          "nombre"
        ],
        "title": "UsuarioCreate",
        "description": "Alta de un usuario nuevo. Solo la puede hacer un admin (ver router)."
      },
      "UsuarioOut": {
        "properties": {
          "id_usuario": {
            "type": "integer",
            "title": "Id Usuario"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "rol": {
            "$ref": "#/components/schemas/RolUsuario"
          },
          "activo": {
            "type": "boolean",
            "title": "Activo"
          }
        },
        "type": "object",
        "required": [
          "id_usuario",
          "email",
          "nombre",
          "rol",
          "activo"
        ],
        "title": "UsuarioOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}